+++ frontend/src/components/Home/Navbar.css
@@ -1,3 +1,4 @@
+/* Navbar Container */
 .navbar {
   display: flex;
   justify-content: space-between;
@@ -11,6 +12,12 @@
   display: flex;
   align-items: center;
   gap: 10px;
+}
+
+.navbar-left h2 {
+  font-size: 18px;
+  color: #ffffff;
+  margin: 0;
 }
 
 .navbar-right {
@@ -51,6 +58,7 @@
   border-radius: 50%;
 }
 
+/* Sidebar Menu for Mobile */
 @media screen and (max-width: 768px) {
   .menu-toggle {
     display: block;
@@ -58,17 +66,45 @@
 
   .navbar-links {
     display: none;
-    flex-direction: column;
-    gap: 10px;
-    position: absolute;
-    top: 60px;
-    right: 20px;
+    position: fixed;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 75%;
     background-color: #101923;
-    padding: 10px 20px;
-    border-radius: 8px;
+    flex-direction: column;
+    padding: 20px;
+    box-shadow: 2px 0 10px rgba(0, 0, 0, 0.5);
+    z-index: 100;
+    transition: transform 0.3s ease-in-out;
+    transform: translateX(-100%);
   }
 
   .navbar-links.open {
     display: flex;
+    transform: translateX(0);
+  }
+
+  .navbar-links a {
+    font-size: 16px;
+    padding: 15px 0;
+    color: #ffffff;
+    border-bottom: 1px solid #223449;
+  }
+
+  .menu-close {
+    display: block;
+    font-size: 24px;
+    color: #ffffff;
+    background: none;
+    border: none;
+    position: absolute;
+    top: 10px;
+    right: 20px;
+    cursor: pointer;
+  }
+
+  .navbar-buttons {
+    display: none;
   }
 }
