+++ frontend/src/components/pages/ActivityPage.css
