+++ frontend/src/components/pages/BusinessNews.css
@@ -1,48 +1,73 @@
 .business-news-container {
     display: flex;
+    flex-direction: column;
     height: 100vh;
-    margin-left: 290px;
-    background-color: #e5e7e6;
+    background-color: #101923; /* Fundal similar cu Invest */
+    color: #ecf0f1; /* Culoare text similară cu Invest */
 }
 
 .news-content {
-    color: #000; /* Aceeași culoare a textului ca navbar-links li a */
     flex-grow: 1;
-    padding: 20px;
+    padding: 2rem 4rem; /* Spațiu consistent cu Invest */
+    overflow-y: auto;
+    color: #ecf0f1; /* Culoare text alb */
+}
+
+.news-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 1.5rem;
+}
+
+.news-header h2 {
+    margin: 0;
+    font-size: 1.5rem;
+    color: #ffffff; /* Titlu alb */
 }
 
 .news-scroll-container {
-    max-height: 80vh;
-    overflow-y: auto;
-    padding-right: 10px;
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    gap: 20px;
+    padding-top: 1rem;
 }
 
 .news-article {
-    margin-bottom: 5px; /* Aceeași marjă ca navbar-links li */
-    padding: 10px; /* Aceeași padding ca navbar-links li a */
-    background-color: #003554; /* Aceeași culoare de fundal ca navbar-links li a */
-    color: #ecf0f1; /* Aceeași culoare a textului ca navbar-links li a */
-    border-radius: 5px; /* Aceeași bordură rotunjită ca navbar-links li a */
-    text-decoration: none; /* Eliminăm sublinierea */
-    display: flex; /* Asigurăm că ocupă întregul spațiu */
-    flex-direction: column; /* Aliniem conținutul pe verticală */
-    align-items: flex-start; /* Aliniem conținutul la stânga */
-    transition: background-color 0.3s; /* Aceeași tranziție ca navbar-links li a */
-    cursor: pointer; /* Adăugăm cursor pointer */
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Adăugăm o umbră pentru consistență */
+    background-color: #223449;
+    color: #ecf0f1;
+    border-radius: 10px;
+    padding: 1rem;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+    transition: transform 0.3s ease, background-color 0.3s ease;
+    cursor: pointer;
+    overflow: hidden;
+    display: flex;
+    flex-direction: column;
 }
 
-.news-article:hover {
-    background-color: #2ec4b6; /* Aceeași culoare de fundal la hover ca navbar-links li a:hover */
-    color: #000; /* Aceeași culoare a textului la hover ca navbar-links li a:hover */
+.news-article img {
+    width: 100%;
+    height: 150px;
+    object-fit: cover;
+    border-radius: 8px;
+    margin-bottom: 1rem;
 }
 
 .news-article h4 {
-    margin: 0 0 10px;
-    text-decoration: none; /* Eliminăm sublinierea */
+    margin: 0;
+    font-size: 1rem;
+    color: #ffffff; /* Titlu alb */
 }
 
 .news-article p {
-    margin: 10px 0 0; /* Adăugăm spațiu între descriere și titlu */
-    text-decoration: none; /* Eliminăm sublinierea */
+    margin-top: 0.5rem;
+    font-size: 0.875rem;
+    color: #90abcb; /* Text descriere albastru deschis */
+}
+
+.news-article:hover {
+    background-color: #358ef3; /* Schimbă fundalul la hover */
+    color: #ffffff;
+    transform: scale(1.05); /* Adaugă efect de hover */
 }
