+++ frontend/src/components/pages/Chatbot.css
@@ -1,94 +1,140 @@
-/* Chatbot.css */
-
-/* General page layout */
-.chatbot-page-container {
-  display: flex;
-  height: 100vh;
-  margin-left: 290px;
-  background-color: #e5e7e6;
+/* Chatbot Container */
+.chatbot-container {
+  position: fixed;
+  bottom: 20px;
+  right: 20px;
+  z-index: 1000;
 }
 
-/* Chatbot content area */
-.chatbot-content {
-  color: #ecf0f1;
-  flex-grow: 1;
-  padding: 20px;
+/* Toggle Button */
+.chatbot-toggle-button {
+  background-color: #358ef3; /* Highlight color from Invest pages */
+  color: white;
+  border: none;
+  border-radius: 50%;
+  width: 60px;
+  height: 60px;
+  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
+  cursor: pointer;
   display: flex;
   align-items: center;
   justify-content: center;
 }
 
-/* Chatbot container */
-.chatbot-container {
+.chat-icon {
+  font-size: 24px;
+}
+
+/* Chatbot Panel */
+.chatbot-panel {
+  position: fixed;
+  bottom: 0;
+  right: 0;
+  width: 400px;
+  height: 600px;
+  background-color: #101923; /* Primary background color */
+  box-shadow: -2px 0px 10px rgba(0, 0, 0, 0.2);
+  transform: translateX(100%);
+  transition: transform 0.3s ease;
   display: flex;
   flex-direction: column;
+  border-radius: 10px 10px 0 0;
+  color: #ffffff; /* Text primary */
+}
+
+.chatbot-panel.open {
+  transform: translateX(0);
+}
+
+/* Chatbot Header */
+.chatbot-header {
+  display: flex;
+  justify-content: space-between;
   align-items: center;
-  justify-content: center;
-  width: 100%;
-  max-width: 600px;
-  margin: 0 auto;
-  padding: 20px;
-  background-color: #e5e7e6;
-  color: #ecf0f1;
+  background-color: #223449; /* Secondary background color */
+  color: #ffffff; /* Primary text */
+  padding: 10px 20px;
+  border-radius: 10px 10px 0 0;
+  font-weight: bold;
+  font-size: 16px;
 }
 
-/* Messages container */
+.close-button {
+  background: none;
+  border: none;
+  color: #ffffff;
+  font-size: 18px;
+  cursor: pointer;
+}
+
+/* Messages Section */
 .chatbot-messages {
-  width: 100%;
-  max-height: 80vh;
+  flex: 1;
   overflow-y: auto;
-  border-radius: 5px;
-  padding: 10px;
-  background-color: #162447;
-  padding-right: 10px; /* Space for scrollbar */
+  padding: 20px;
+  background-color: #101923; /* Primary background color */
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
 }
 
-/* Individual messages */
 .chatbot-message {
-  padding: 8px 10px;
-  margin: 5px 0;
-  border-radius: 5px;
   max-width: 80%;
-  color: #ecf0f1;
+  padding: 10px;
+  border-radius: 8px;
+  font-size: 14px;
 }
 
 .chatbot-message.user {
-  background-color: #007bff;
-  align-self: flex-end;
+  background-color: #358ef3; /* Highlight color */
   color: white;
+  align-self: flex-end;
 }
 
 .chatbot-message.bot {
-  background-color: #f8d7da;
+  background-color: #223449; /* Secondary background */
+  color: #90abcb; /* Secondary text color */
   align-self: flex-start;
-  color: black;
 }
 
-/* Input container */
+/* Input Container */
 .chatbot-input-container {
   display: flex;
-  width: 100%;
-  margin-top: 10px;
+  padding: 10px;
+  background-color: #223449; /* Secondary background */
+  border-top: 1px solid #314b68; /* Subtle border */
 }
 
 .chatbot-input {
-  flex-grow: 1;
+  flex: 1;
   padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
+  border: 1px solid #314b68;
+  border-radius: 20px;
+  outline: none;
+  font-size: 14px;
+  margin-right: 8px;
+  background-color: #101923; /* Primary background color */
+  color: #ffffff; /* Primary text color */
 }
 
-.chatbot-button {
-  padding: 10px 20px;
-  margin-left: 10px;
+.chatbot-send-button {
+  background-color: #358ef3; /* Highlight color */
   border: none;
-  border-radius: 5px;
-  background-color: #007bff;
   color: white;
+  border-radius: 50%;
+  width: 40px;
+  height: 40px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   cursor: pointer;
-  transition: background-color 0.3s;
 }
 
-.chatbot-button:hover {
-  background-color: #00a706;
+/* Responsive Design */
+@media screen and (max-width: 600px) {
+  .chatbot-panel {
+    width: 100%;
+    height: 80%;
+    border-radius: 10px 10px 0 0;
+  }
 }
