+++ frontend/src/components/pages/Invest1.css
