+++ frontend/src/components/pages/Products.css
@@ -2,78 +2,81 @@
 .stock-list-container {
     display: flex;
     flex-direction: column;
-    min-height: 100vh; /* Full height */
-    margin-left: 290px; /* Leave space for the navbar */
-    background-color: #e5e7e6; /* Soft background color */
-    padding: 20px;
+    min-height: 100vh;
+    background-color: #101923;
+    color: #e5e7e6;
+    padding: 0; /* Removed padding to allow full width */
+    width: 100%;
+    margin: 0; /* Removed any unintended margin */
 }
 
 /* Main content area */
 .stock-list-content {
-    color: #2c3e50; /* Dark text color for readability */
     flex-grow: 1;
-    padding: 20px;
-    border-radius: 10px;
-    background: #e5e7e6; /* Card background for contrast */
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1); /* Subtle shadow */
+    padding: 0; /* Removed padding */
+    border-radius: 0; /* Removed border radius */
+    background: #101923;
+    width: 100%;
 }
 
 /* Stock list styling */
 .stock-list {
-    display: grid; /* Arrange in grid */
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr)); /* Flexible grid */
-    gap: 20px; /* Space between cards */
-    max-height: 80vh; /* Prevent overflow */
-    overflow-y: auto; /* Enable scrolling */
-    padding-right: 10px; /* Space for scrollbar */
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 10px;
+    max-height: 80vh;
+    overflow-y: auto;
+    width: 100%;
+    margin: 0;
 }
 
 /* Stock item (individual card) */
 .stock-item {
-    background-color: #ffffff; /* White card background */
-    border-radius: 10px; /* Rounded corners */
-    padding: 20px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Soft shadow */
-    transition: transform 0.3s ease, box-shadow 0.3s ease; /* Hover effects */
+    background-color: #223449;
+    border-radius: 10px;
+    padding: 10px;
+    transition: background-color 0.3s ease, box-shadow 0.3s ease;
     display: flex;
     flex-direction: column;
     justify-content: space-between;
+    border: 1px solid #3a3d46;
+    color: #e5e7e6;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
 }
 
 .stock-item:hover {
-    transform: translateY(-5px); /* Lift the card */
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15); /* Stronger shadow */
-    background-color: #f0f4ff; /* Light blue hover background */
+    background-color: #358ef3;
+    color: #ffffff;
+    transform: translateY(-5px);
+    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.2);
 }
 
 /* Header inside each card */
 .stock-header {
     display: flex;
     align-items: center;
-    margin-bottom: 15px;
+    margin-bottom: 5px;
 }
 
 .stock-header h2 {
-    font-size: 20px;
-    color: #003554; /* Blue header */
+    font-size: 18px;
     margin: 0;
+    font-weight: bold;
 }
 
 .logo-placeholder {
     width: 40px;
     height: 40px;
-    background-color: #e0e0e0; /* Placeholder for logo */
+    background-color: #3a3d46;
     border-radius: 50%;
-    margin-right: 15px;
+    margin-right: 10px;
 }
 
 /* Stock details */
 .stock-details {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 15px;
+    margin-bottom: 10px;
     font-size: 16px;
-    color: #555; /* Subtle text color */
+    color: #9ea2a8;
 }
 
 .stock-label {
@@ -81,17 +84,17 @@
 }
 
 .stock-value {
-    color: #27ae60; /* Green for positive numbers */
+    color: #4caf50; /* Green for positive numbers */
 }
 
 /* Buttons inside each card */
 .stock-buttons {
     display: flex;
-    gap: 10px;
+    gap: 5px;
 }
 
 .stock-buttons button {
-    padding: 10px 20px;
+    padding: 8px 15px;
     font-size: 14px;
     font-weight: bold;
     border: none;
@@ -101,58 +104,67 @@
 }
 
 .stock-buttons .buy-button {
-    background-color: #1abc9c; /* Teal for buy */
+    background-color: #1abc9c;
     color: white;
 }
 
 .stock-buttons .buy-button:hover {
-    background-color: #16a085; /* Darker teal */
-    transform: scale(1.05); /* Slight enlargement */
+    background-color: #16a085;
+    transform: scale(1.05);
 }
 
 .stock-buttons .sell-button {
-    background-color: #e74c3c; /* Red for sell */
+    background-color: #e74c3c;
     color: white;
 }
 
 .stock-buttons .sell-button:hover {
-    background-color: #c0392b; /* Darker red */
-    transform: scale(1.05); /* Slight enlargement */
+    background-color: #c0392b;
+    transform: scale(1.05);
 }
 
 /* Input box */
 .stock-input-box {
-    background-color: #f8f9fa; /* Light gray */
-    color: #34495e; /* Dark gray text */
-    border: 1px solid #bdc3c7; /* Light border */
+    background-color: #223449;
+    color: #e5e7e6;
+    border: 1px solid #3a3d46;
     border-radius: 8px;
-    padding: 15px;
-    margin-bottom: 20px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Subtle shadow */
+    padding: 10px;
+    margin-bottom: 15px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
 
 .stock-input-box input {
     width: 100%;
-    padding: 12px;
+    padding: 10px;
     font-size: 16px;
-    border: 2px solid #3498db; /* Blue border */
+    border: 2px solid #3498db;
     border-radius: 8px;
-    margin-bottom: 10px;
+    margin-bottom: 8px;
     outline: none;
+    background-color: #1c1c24;
+    color: #e5e7e6;
     transition: border-color 0.2s ease, box-shadow 0.2s ease;
 }
 
 .stock-input-box input:focus {
-    border-color: #2980b9; /* Darker blue */
-    box-shadow: 0 0 8px rgba(41, 128, 185, 0.5); /* Glow effect */
+    border-color: #2980b9;
+    box-shadow: 0 0 8px rgba(41, 128, 185, 0.5);
+}
+
+.stock-input-box .buttons-container {
+    display: flex;
+    justify-content: space-between; /* Separated buttons */
+    gap: 10px;
 }
 
 .stock-input-box button {
-    padding: 10px 18px;
+    flex: 1; /* Equal width for buttons */
+    padding: 8px 15px;
     font-size: 14px;
-    font-weight: 600; /* Bold text */
+    font-weight: 600;
     color: white;
-    background-color: #003554; /* Blue button */
+    background-color: #358ef3;
     border: none;
     border-radius: 5px;
     cursor: pointer;
@@ -160,20 +172,50 @@
 }
 
 .stock-input-box button:hover {
-    background-color: #2980b9; /* Darker blue */
-    transform: scale(1.05); /* Slight enlargement */
+    background-color: #003554;
+    transform: scale(1.05);
 }
 
 /* Scrollbar for stock list */
 .stock-list::-webkit-scrollbar {
-    width: 8px;
+    width: 6px;
 }
 
 .stock-list::-webkit-scrollbar-thumb {
-    background-color: #bdc3c7; /* Light gray scrollbar */
+    background-color: #555;
     border-radius: 10px;
 }
 
 .stock-list::-webkit-scrollbar-thumb:hover {
-    background-color: #7f8c8d; /* Darker gray on hover */
+    background-color: #7f8c8d;
+}
+
+/* Responsive Styling */
+@media (max-width: 768px) {
+    .stock-list {
+        grid-template-columns: repeat(auto-fill, minmax(200px, 1fr)); /* Smaller cards for smaller screens */
+    }
+
+    .stock-input-box {
+        padding: 8px; /* Reduced padding */
+    }
+
+    .stock-buttons button {
+        font-size: 12px; /* Smaller font for buttons */
+        padding: 6px 10px;
+    }
+}
+
+@media (max-width: 480px) {
+    .stock-list {
+        grid-template-columns: 1fr; /* Single column on very small screens */
+    }
+
+    .stock-input-box .buttons-container {
+        flex-direction: column; /* Stack buttons vertically */
+    }
+
+    .stock-input-box button {
+        margin-bottom: 8px;
+    }
 }
