+++ frontend/src/components/pages/Transactions.css
@@ -1,100 +1,61 @@
+/* Transactions.css */
 .transaction-list-container {
-    display: flex;
-    height: 100vh;
-    margin-left: 290px;
-    background-color: #E5E7E6;
-}
-
-.transaction-list-content {
-    color: #000;
-    flex-grow: 1;
     padding: 20px;
-}
-
-.transaction-list {
-    max-height: 80vh;
-    overflow-y: auto;
-    padding-right: 10px;
-}
-
-.transaction-item {
-    margin-bottom: 5px;
-    padding: 10px;
+    background-color: #101923;
+    color: #ffffff;
+  }
+  
+  .transaction-list-header {
+    font-size: 28px;
+    font-weight: bold;
+    margin-bottom: 30px;
+  }
+  
+  .transaction-list {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 20px;
+  }
+  
+  .transaction-item {
     background-color: #162447;
-    color: #ecf0f1;
-    border-radius: 5px;
-    text-decoration: none;
+    border-radius: 8px;
+    padding: 20px;
     display: flex;
     flex-direction: column;
-    align-items: flex-start;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    transition: background-color 0.3s;
-    cursor: default;
-}
-
-.transaction-header {
-    display: flex;
-    align-items: center;
-    margin-bottom: 10px;
-}
-
-.logo-placeholder {
-    width: 40px;
-    height: 40px;
-    background-color: #ccc;
-    margin-right: 10px;
-    border-radius: 5px;
-}
-
-.transaction-header h2 {
-    margin: 0;
-}
-
-.transaction-details {
-    display: flex;
     justify-content: space-between;
-    width: 100%;
-    margin-bottom: 5px;
-}
-
-.transaction-label {
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
+    transition: transform 0.2s ease, box-shadow 0.2s ease;
+  }
+  
+  .transaction-item:hover {
+    transform: scale(1.05);
+    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.3);
+  }
+  
+  .transaction-item h3 {
+    font-size: 16px;
     font-weight: bold;
-    margin-right: 10px;
-}
-
-.transaction-value {
-    font-weight: normal;
-}
-
-.transaction-buttons {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    margin-top: 10px;
-}
-
-.transaction-buttons button {
-    padding: 5px 10px;
-    border: none;
-    border-radius: 3px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.transaction-buttons .approve-button {
-    background-color: #007bff;
-    color: white;
-}
-
-.transaction-buttons .approve-button:hover {
-    background-color: #00a706;
-}
-
-.transaction-buttons .reject-button {
-    background-color: #007bff;
-    color: white;
-}
-
-.transaction-buttons .reject-button:hover {
-    background-color: #b7000f;
-}
+    margin-bottom: 10px;
+  }
+  
+  .transaction-item .transaction-details {
+    font-size: 14px;
+    line-height: 1.5;
+  }
+  
+  .transaction-item .transaction-details span {
+    display: block;
+    margin-bottom: 5px;
+  }
+  
+  .transaction-item .transaction-price {
+    font-weight: bold;
+    color: #28a745;
+  }
+  
+  .transaction-item .transaction-date {
+    font-size: 12px;
+    color: #999;
+  }
+
