--- frontend/src/pages/Balance.css
@@ -1,133 +1,0 @@
-.balance-container {
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-    margin-left: 290px;
-    background-color: #E5E7e6;
-    color: #333;
-    font-family: 'Arial', sans-serif;
-    padding: 20px;
-    gap: 30px;
-}
-
-/* Header pentru Balance */
-.balance-header {
-    background: #003554;
-    padding: 25px;
-    border-radius: 15px;
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.1);
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    color: white;
-}
-
-.balance-header h2 {
-    margin: 0;
-    font-size: 28px;
-    font-weight: bold;
-}
-
-.balance-header p {
-    font-size: 26px;
-    font-weight: bold;
-    color: #fff;
-    background-color: rgba(0, 0, 0, 0.2);
-    padding: 10px 20px;
-    border-radius: 10px;
-}
-
-/* Stocuri Header */
-.stocks-header {
-    font-size: 26px;
-    font-weight: bold;
-    color: #4a4a4a;
-    margin-bottom: 15px;
-}
-
-.stock-list {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-    gap: 20px;
-    padding: 0;
-    list-style: none;
-    margin: 0;
-}
-
-/* Card individual pentru fiecare stoc */
-.stock-item {
-    background: white;
-    border-radius: 15px;
-    padding: 20px;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: space-between;
-    box-shadow: 0 8px 15px rgba(0, 0, 0, 0.1);
-    transition: transform 0.3s ease, box-shadow 0.3s ease;
-}
-
-.stock-item:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 12px 25px rgba(0, 0, 0, 0.15);
-}
-
-.stock-symbol {
-    font-size: 20px;
-    font-weight: bold;
-    color: #333;
-}
-
-.stock-quantity {
-    font-size: 18px;
-    margin-top: 10px;
-    color: #003554;
-}
-
-/* Secțiunea pentru acțiuni */
-.balance-actions {
-    background: #003554;
-    padding: 25px;
-    border-radius: 15px;
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.1);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 20px;
-    color: white;
-}
-
-.balance-actions input {
-    padding: 15px;
-    border-radius: 10px;
-    border: 2px solid white;
-    background-color: #f5f5f5;
-    color: #333;
-    width: 100%;
-    max-width: 350px;
-    font-size: 18px;
-    transition: border-color 0.3s ease;
-}
-
-.balance-actions input:focus {
-    border-color: #2575fc;
-    outline: none;
-}
-
-.balance-actions button {
-    padding: 15px 30px;
-    border-radius: 10px;
-    border: none;
-    background-color: white;
-    color: #2575fc;
-    font-weight: bold;
-    cursor: pointer;
-    transition: background-color 0.3s ease, transform 0.3s ease;
-    font-size: 18px;
-}
-
-.balance-actions button:hover {
-    background-color: #003554;
-    color: white;
-    transform: scale(1.05);
-}
