--- frontend/src/pages/BusinessNews.css
@@ -1,73 +1,0 @@
-.business-news-container {
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-    background-color: #101923; /* Fundal similar cu Invest */
-    color: #ecf0f1; /* Culoare text similară cu Invest */
-}
-
-.news-content {
-    flex-grow: 1;
-    padding: 2rem 4rem; /* Spațiu consistent cu Invest */
-    overflow-y: auto;
-    color: #ecf0f1; /* Culoare text alb */
-}
-
-.news-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 1.5rem;
-}
-
-.news-header h2 {
-    margin: 0;
-    font-size: 1.5rem;
-    color: #ffffff; /* Titlu alb */
-}
-
-.news-scroll-container {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-    gap: 20px;
-    padding-top: 1rem;
-}
-
-.news-article {
-    background-color: #223449;
-    color: #ecf0f1;
-    border-radius: 10px;
-    padding: 1rem;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-    transition: transform 0.3s ease, background-color 0.3s ease;
-    cursor: pointer;
-    overflow: hidden;
-    display: flex;
-    flex-direction: column;
-}
-
-.news-article img {
-    width: 100%;
-    height: 150px;
-    object-fit: cover;
-    border-radius: 8px;
-    margin-bottom: 1rem;
-}
-
-.news-article h4 {
-    margin: 0;
-    font-size: 1rem;
-    color: #ffffff; /* Titlu alb */
-}
-
-.news-article p {
-    margin-top: 0.5rem;
-    font-size: 0.875rem;
-    color: #90abcb; /* Text descriere albastru deschis */
-}
-
-.news-article:hover {
-    background-color: #358ef3; /* Schimbă fundalul la hover */
-    color: #ffffff;
-    transform: scale(1.05); /* Adaugă efect de hover */
-}
