--- frontend/src/pages/Chatbot.css
@@ -1,140 +1,0 @@
-/* Chatbot Container */
-.chatbot-container {
-  position: fixed;
-  bottom: 20px;
-  right: 20px;
-  z-index: 1000;
-}
-
-/* Toggle Button */
-.chatbot-toggle-button {
-  background-color: #358ef3; /* Highlight color from Invest pages */
-  color: white;
-  border: none;
-  border-radius: 50%;
-  width: 60px;
-  height: 60px;
-  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.chat-icon {
-  font-size: 24px;
-}
-
-/* Chatbot Panel */
-.chatbot-panel {
-  position: fixed;
-  bottom: 0;
-  right: 0;
-  width: 400px;
-  height: 600px;
-  background-color: #101923; /* Primary background color */
-  box-shadow: -2px 0px 10px rgba(0, 0, 0, 0.2);
-  transform: translateX(100%);
-  transition: transform 0.3s ease;
-  display: flex;
-  flex-direction: column;
-  border-radius: 10px 10px 0 0;
-  color: #ffffff; /* Text primary */
-}
-
-.chatbot-panel.open {
-  transform: translateX(0);
-}
-
-/* Chatbot Header */
-.chatbot-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #223449; /* Secondary background color */
-  color: #ffffff; /* Primary text */
-  padding: 10px 20px;
-  border-radius: 10px 10px 0 0;
-  font-weight: bold;
-  font-size: 16px;
-}
-
-.close-button {
-  background: none;
-  border: none;
-  color: #ffffff;
-  font-size: 18px;
-  cursor: pointer;
-}
-
-/* Messages Section */
-.chatbot-messages {
-  flex: 1;
-  overflow-y: auto;
-  padding: 20px;
-  background-color: #101923; /* Primary background color */
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.chatbot-message {
-  max-width: 80%;
-  padding: 10px;
-  border-radius: 8px;
-  font-size: 14px;
-}
-
-.chatbot-message.user {
-  background-color: #358ef3; /* Highlight color */
-  color: white;
-  align-self: flex-end;
-}
-
-.chatbot-message.bot {
-  background-color: #223449; /* Secondary background */
-  color: #90abcb; /* Secondary text color */
-  align-self: flex-start;
-}
-
-/* Input Container */
-.chatbot-input-container {
-  display: flex;
-  padding: 10px;
-  background-color: #223449; /* Secondary background */
-  border-top: 1px solid #314b68; /* Subtle border */
-}
-
-.chatbot-input {
-  flex: 1;
-  padding: 10px;
-  border: 1px solid #314b68;
-  border-radius: 20px;
-  outline: none;
-  font-size: 14px;
-  margin-right: 8px;
-  background-color: #101923; /* Primary background color */
-  color: #ffffff; /* Primary text color */
-}
-
-.chatbot-send-button {
-  background-color: #358ef3; /* Highlight color */
-  border: none;
-  color: white;
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-}
-
-/* Responsive Design */
-@media screen and (max-width: 600px) {
-  .chatbot-panel {
-    width: 100%;
-    height: 80%;
-    border-radius: 10px 10px 0 0;
-  }
-}
