--- frontend/src/pages/Home.css
@@ -1,105 +1,0 @@
-/* General container styling */
-.home-container {
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-    background-color: #101923;
-    color: #e5e7e6;
-  }
-  
-  /* Content styling */
-  .home-content {
-    flex-grow: 1;
-    padding: 20px 40px;
-  }
-  
-  /* Header styling */
-  .home-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 20px;
-  }
-  
-  .home-header h2 {
-    margin: 0;
-    font-size: 24px;
-    color: #ffffff;
-  }
-  
-  .header-icons i {
-    margin-right: 15px;
-    font-size: 20px;
-    color: #ffffff;
-    cursor: pointer;
-  }
-  
-  /* Card styling */
-  .cards {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 20px;
-  }
-  
-  .card {
-    flex: 1 1 calc(33.333% - 20px);
-    max-width: 300px;
-    background-color: #223449;
-    color: #e5e7e6;
-    border-radius: 10px;
-    padding: 15px;
-    transition: background-color 0.3s ease;
-  }
-  
-  .card:hover {
-    background-color: #358ef3;
-    color: #ffffff;
-  }
-  
-  .card-title {
-    font-size: 18px;
-    font-weight: bold;
-  }
-  
-  .card-value {
-    font-size: 16px;
-    margin: 10px 0;
-  }
-  
-  .card-change.green {
-    color: #4caf50;
-  }
-  
-  .card-change.red {
-    color: #f44336;
-  }
-  
-  /* Chart container */
-  .chart-container {
-    margin-top: 30px;
-    background-color: #223449;
-    border-radius: 10px;
-    padding: 20px;
-    color: #e5e7e6;
-  }
-  
-  /* Chatbot button */
-  .chatbot-icon {
-    position: fixed;
-    bottom: 20px;
-    right: 20px;
-    background-color: #358ef3;
-    color: #ffffff;
-    border: none;
-    border-radius: 50%;
-    width: 50px;
-    height: 50px;
-    font-size: 24px;
-    cursor: pointer;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  }
-  
-  .chatbot-icon:hover {
-    background-color: #003554;
-  }
-
