--- frontend/src/pages/Profile.css
@@ -1,110 +1,0 @@
-/* Profile.css */
-
-/* General container styling */
-.profile-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    background-color: #e5e7e6;
-    min-height: 100vh;
-    padding: 20px;
-}
-
-/* Profile card styling */
-.profile-card-custom {
-    background: #003554;
-    border-radius: 10px;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
-    padding: 30px;
-    width: 90%;
-    max-width: 600px;
-    text-align: center;
-}
-
-.profile-card-custom h2 {
-    color: #e5e7e6;
-    font-size: 24px;
-    margin-bottom: 20px;
-}
-
-/* Profile field styling */
-.profile-field-custom {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    margin-bottom: 15px;
-}
-
-.profile-field-custom label {
-    font-size: 14px;
-    font-weight: bold;
-    color: #e5e7e6;
-    margin-bottom: 5px;
-}
-
-.profile-field-custom p {
-    font-size: 16px;
-    color: #e5e7e6;
-    margin: 0;
-}
-
-.profile-field-custom input {
-    width: 100%;
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ced4da;
-    border-radius: 5px;
-    background-color: #f8f9fa;
-    transition: border-color 0.2s;
-}
-
-.profile-field-custom input:focus {
-    outline: none;
-    border-color: #007bff;
-    background-color: #ffffff;
-    box-shadow: 0 0 5px rgba(0, 123, 255, 0.25);
-}
-
-/* Buttons */
-.chart-buttons {
-    display: flex;
-    justify-content: center;
-    margin-top: 20px;
-}
-
-.chart-buttons button {
-    padding: 10px 20px;
-    font-size: 16px;
-    font-weight: bold;
-    color: #ffffff;
-    background-color: #007bff;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.chart-buttons button:hover {
-    background-color: #0056b3;
-}
-
-.chart-buttons button:active {
-    background-color: #004085;
-}
-
-/* Responsive design */
-@media (max-width: 480px) {
-    .profile-card-custom {
-        padding: 20px;
-    }
-
-    .profile-field-custom input {
-        font-size: 14px;
-    }
-
-    .chart-buttons button {
-        font-size: 14px;
-        padding: 8px 16px;
-    }
-}
