--- frontend/src/pages/Transactions.css
@@ -1,61 +1,0 @@
-/* Transactions.css */
-.transaction-list-container {
-    padding: 20px;
-    background-color: #101923;
-    color: #ffffff;
-  }
-  
-  .transaction-list-header {
-    font-size: 28px;
-    font-weight: bold;
-    margin-bottom: 30px;
-  }
-  
-  .transaction-list {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-    gap: 20px;
-  }
-  
-  .transaction-item {
-    background-color: #162447;
-    border-radius: 8px;
-    padding: 20px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
-    transition: transform 0.2s ease, box-shadow 0.2s ease;
-  }
-  
-  .transaction-item:hover {
-    transform: scale(1.05);
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.3);
-  }
-  
-  .transaction-item h3 {
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 10px;
-  }
-  
-  .transaction-item .transaction-details {
-    font-size: 14px;
-    line-height: 1.5;
-  }
-  
-  .transaction-item .transaction-details span {
-    display: block;
-    margin-bottom: 5px;
-  }
-  
-  .transaction-item .transaction-price {
-    font-weight: bold;
-    color: #28a745;
-  }
-  
-  .transaction-item .transaction-date {
-    font-size: 12px;
-    color: #999;
-  }
-
